+++ src/Options/XAxis/Crosshairs.php
@@ -16,7 +16,7 @@ class Crosshairs extends AxisCrosshairsAbstract
         parent::__construct($options);
     }
 
-    public function width(int $value): static
+    public function width(float $value): static
     {
         return $this->setOption('width', $value);
     }
